Added support for alternative blocks in blocksets.
Added skinmanager framework.
Widget and Item ID's are now 64 bit.
Added RemoveResDB() to WgResDB.
XML-file included from other XML-file can now reference material already
stored in parents and siblings resDB's.
VectorGlyphs now supports size offset.

+++ source/wg_text.cpp
@@ -932,7 +932,7 @@ WgChar * WgText::parseScaledValue( Sint64 value, Uint32 scale, const WgValueForm
 
 	while( intPart > 0 )
 	{
-		temp2[n++] = (intPart % 10) + 0x30;
+		temp2[n++] = (Uint16) ((intPart % 10) + 0x30);
 		intPart /= 10;
 	}
